+++ src/plugin/controller.ts
@@ -1,4 +1,5 @@
 import {postMessage} from "./postMessage";
+import {loopChildTextNodes, loopChildFrameNodes} from "./utils";
 
 function main() {
     figma.showUI(__html__, {width: 240, height: 320});
@@ -89,18 +90,6 @@ async function fillCards(data) {
     }
 }
 
-function shouldReplaceText(node) {
-    // Return TRUE when the node type == "TEXT" AND when the node name includes "#"
-    // Otherwise return FALSE
-    return node.type === "TEXT" && node.name.includes("#");
-}
-
-function isImage(node) {
-    // Return TRUE when the node type == "FRAME" AND when the node name includes "#"
-    // Otherwise return FALSE
-    return node.name.includes("Image");
-}
-
 const gatherValue = (name, row) => {
     const layerName = name.replace("#", "").split(".");
     let value = layerName.reduce(function (obj, prop) {
@@ -128,37 +117,6 @@ async function replaceText(node, content) {
      * Once fonts are loaded we can change the text
      */
     node.characters = String(content);
-}
-
-function loopChildTextNodes(nodes, row) {
-    let textMatches = [];
-    for (let i = 0; i < nodes.length; i++) {
-        const node = nodes[i];
-        if (shouldReplaceText(node)) {
-            textMatches.push([node, row]);
-        }
-        if (node.children) {
-            const nextTextMatches = loopChildTextNodes(node.children, row);
-            textMatches = [...textMatches, ...nextTextMatches];
-        }
-    }
-    return textMatches;
-}
-
-// Loop through child nodes, check whether their name includes "Image" and push their ids in an array
-function loopChildFrameNodes(nodes, row) {
-    let imageFrames = [];
-    for (let i = 0; i < nodes.length; i++) {
-        const node = nodes[i];
-        if (isImage(node)) {
-            imageFrames.push([node.id, node.name, row]);
-        }
-        if (node.children) {
-            const nextImageFrames = loopChildFrameNodes(node.children, row);
-            imageFrames = [...imageFrames, ...nextImageFrames];
-        }
-    }
-    return imageFrames;
 }
 
 main();
